implement array_ops (#27936)
